cleaned up tets
