+++ tests/test_write.py
@@ -1,5 +1,5 @@
-from typing import List
 import pytest
+from tests.helpers import compare_rows_helper
 
 from doltcli import (
     write_rows,
@@ -23,22 +23,6 @@ TEST_COLUMNS = {
     'date_of_death': ['1877-01-01', '', '']
 }
 
-
-def compare_rows_helper(expected: List[dict], actual: List[dict]):
-    assert len(expected) == len(actual), f'Unequal row counts: {len(expected)} != {len(actual)}'
-    errors = []
-    for l, r in zip(expected, actual):
-        l_cols, r_cols = set(l.keys()), set(r.keys())
-        assert l_cols == r_cols, f'Unequal sets of columns: {l_cols} != {r_cols}'
-        for col in l_cols:
-            l_val, r_val = l[col], r[col]
-            if col.startswith('date'):
-                l_val, r_val = l_val[:10], r_val[:10]
-            if l_val != r_val and not (l_val is None and r_val == ''):
-                errors.append(f'{col}: {l_val} != {r_val}')
-
-    error_str = '\n'.join(errors)
-    assert not errors, f'Failed with the following unequal columns:\n{error_str}'
 
 def test_write_rows(init_empty_test_repo):
     dolt = init_empty_test_repo
